+++ assets/src/scripts/components/dailyValueHydrograph/selectors/labels.js
@@ -1,12 +1,12 @@
 import {createSelector} from 'reselect';
 
-import {getCurrentObservationsTimeSeries} from '../../../selectors/observations-selector';
+import {getCurrentDVTimeSeries} from '../../../selectors/observations-selector';
 
 /*
  * Returns a selector function that returns a string that can be used as the title of the graph
  */
 export const getCurrentTimeSeriesTitle = createSelector(
-    getCurrentObservationsTimeSeries,
+    getCurrentDVTimeSeries,
     (timeSeries) => {
         return timeSeries && timeSeries.properties ?
             timeSeries.properties.observedPropertyName : '';
@@ -17,7 +17,7 @@ export const getCurrentTimeSeriesTitle = createSelector(
  * Returns a selector function that returns a string that can be used as the description of the graph
  */
 export const getCurrentTimeSeriesDescription = createSelector(
-    getCurrentObservationsTimeSeries,
+    getCurrentDVTimeSeries,
     (timeSeries) => {
         return timeSeries && timeSeries.properties ?
             `${timeSeries.properties.observedPropertyName} for ${timeSeries.properties.samplingFeatureName}` : '';
@@ -28,7 +28,7 @@ export const getCurrentTimeSeriesDescription = createSelector(
  * Returns a selector function that returns a string that can be used as the label of the y axis graph
  */
 export const getCurrentTimeSeriesYTitle = createSelector(
-    getCurrentObservationsTimeSeries,
+    getCurrentDVTimeSeries,
     (timeSeries) => {
         return timeSeries && timeSeries.properties ?
             `${timeSeries.properties.observedPropertyName}, ${timeSeries.properties.unitOfMeasureName}` : '';