+++ assets/tests/scripts/global-config.js
@@ -2,6 +2,8 @@ var CONFIG = {
     TIMESERIES_AUDIO_ENABLED: true,
     MULTIPLE_TIME_SERIES_METADATA_SELECTOR_ENABLED: true,
     FIM_GIS_ENDPOINT: 'https:/fakelegendservice.com',
+    NETWORK_ENDPOINT: 'https://fakeogcservice.com/observations/collections',
+    OBSERVATIONS_ENDPOINT: 'https://fakeogcservice.com/observations/collections/monitoring-locations',
     uvPeriodOfRecord: {
         '00010': {
             begin_date: '01-02-2001',